+++ processor/src/main/java/react4j/processor/ViewType.java
@@ -8,5 +8,6 @@ enum ViewType
 {
   STATEFUL,
   TRACKING,
-  MAYBE_TRACKING
+  MAYBE_TRACKING,
+  NO_RENDER
 }