+++ processor/src/test/fixtures/bad_input/com/example/render/UnexpectedRenderComponent.java
@@ -4,7 +4,7 @@ import react4j.ReactNode;
 import react4j.annotations.Render;
 import react4j.annotations.View;
 
-@View( requireRender = false )
+@View( type = View.Type.NO_RENDER )
 abstract class UnexpectedRenderComponent
 {
   @Render