+++ processor/src/test/fixtures/input/com/example/render/NoRenderWithPostConstructView.java
@@ -3,7 +3,7 @@ package com.example.render;
 import arez.annotations.PostConstruct;
 import react4j.annotations.View;
 
-@View( requireRender = false )
+@View( type = View.Type.NO_RENDER )
 abstract class NoRenderWithPostConstructView
 {
   @PostConstruct