+++ processor/src/test/fixtures/input/com/example/render/NoRenderWithPostMountView.java
@@ -3,7 +3,7 @@ package com.example.render;
 import react4j.annotations.PostMount;
 import react4j.annotations.View;
 
-@View( requireRender = false )
+@View( type = View.Type.NO_RENDER )
 abstract class NoRenderWithPostMountView
 {
   @PostMount