+++ processor/src/test/fixtures/input/com/example/render/NoRenderWithPostUpdateView.java
@@ -3,7 +3,7 @@ package com.example.render;
 import react4j.annotations.PostUpdate;
 import react4j.annotations.View;
 
-@View( requireRender = false )
+@View( type = View.Type.NO_RENDER )
 abstract class NoRenderWithPostUpdateView
 {
   @PostUpdate