+++ processor/src/test/fixtures/input/com/example/render/NoRenderWithPreUpdateView.java
@@ -3,7 +3,7 @@ package com.example.render;
 import react4j.annotations.PreUpdate;
 import react4j.annotations.View;
 
-@View( requireRender = false )
+@View( type = View.Type.NO_RENDER )
 abstract class NoRenderWithPreUpdateView
 {
   @PreUpdate