+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Ensure that the static methods for array inputs are generated as varargs methods. The equivalent instance methods were converted to vararg methods in `0.170`.
+* Change the way "no render" views are specified from `@View(requireRender=false)` to `@View(type=NO_RENDER)`.
 
 ### [v0.171](https://github.com/react4j/react4j/tree/v0.171) (2020-07-21) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.170...v0.171) · [API Differences](https://react4j.github.io/api-diff?key=react4j&old=0.170&new=0.171)